Update RPL FSPs to 4221.00
Update the FSPs for RPL to:

- Raptor Lake P/H/PX/U
  - Version: 4221.00
  - ID: 781590
- Raptor Lake S/HX
  - Version: 4221.00
  - ID: 781583

This is the version Intel used for the most recent upstream coreboot
changes.

As part of this, FSPs are now moved to `fsp/<soc>/<version>` instead of
being in an arbitrary model's directory. All symlinks have been updated
to point to the new locations.

Ref: Intel Software Kit ID 686654
Ref: https://review.coreboot.org/c/coreboot/+/76147
Signed-off-by: Tim Crawford <tcrawford@system76.com>